decompose learner
