+++ nn/src/learner/__init__.py
@@ -1,1 +1,2 @@
 from .baselearner import *
+from .supervisedlearner import *
