+++ nn/src/learner/baselearner.py
@@ -1,121 +1,39 @@
 import torch
-from datasets import *
 from utils.typing import *
-from metrics import Metric
-from utils import (
-    vprint,
-    get_device,
-    move_to,
-    save_model,
-    load_model,
-    detach,
-    AverageValueMeter,
-)
-from test import evaluate
 from utils.logger import TensorboardLogger
-from scheduler import *
-
-import numpy as np
-from torch.optim import Optimizer
-from torch.utils.data import DataLoader
-from torch.nn import Module
-from torch import device
-from torch.cuda.amp import GradScaler, autocast
-
-from tqdm import tqdm
-
 from pathlib import Path
+from tqdm import tqdm
+from utils import vprint, AverageValueMeter, detach, move_to
+from torch.cuda.amp import GradScaler, autocast
+from torch.utils.data import DataLoader
 
 
 class BaseLearner:
     def __init__(
         self,
+        exp_id: str,
+        opt: Any,
+        save_dir: str,
         train_data: DataLoader,
         val_data: DataLoader,
+        device: torch.device,
         model: Module,
-        criterion: Module,
-        optimizer: Optimizer,
-        save_dir: str,
+        optimizer: torch.optim.Optimizer,
+        criterion: Optional[Module] = None,
+        verbose: bool = True,
     ):
         self.train_data, self.val_data = train_data, val_data
         self.model, self.criterion, self.optimizer = model, criterion, optimizer
-        self.save_dir = Path(save_dir)
-
-
-class SupervisedLearner(BaseLearner):
-    def __init__(
-        self,
-        train_data: DataLoader,
-        val_data: DataLoader,
-        model: Module,
-        criterion: Module,
-        optimizer: Optimizer,
-        device: device = get_device(),
-        load_path: Optional[str] = None,
-        save_dir: str = "./",
-    ):
-        super().__init__(train_data, val_data, model, criterion, optimizer, save_dir)
-        self.device = device
-
-        if load_path is not None:
-            load_model(self.model, load_path)
+        self.save_dir = Path(save_dir) / exp_id
         self.tsboard = TensorboardLogger(path=self.save_dir)
-        self.verbose = True
+        self.opt = opt
+        self.device = device
+        self.scaler = GradScaler(enabled=False)
         self.max_grad_norm = 1.0
-
-    def fit(
-        self,
-        n_epochs: int,
-        metrics: Dict[str, Metric] = None,
-        log_step=1,
-        val_step=1,
-        fp16: bool = False,
-        verbose: bool = True,
-    ) -> float:
-        self.fp16 = fp16
-        self.debug = False
         self.verbose = verbose
-        self.metric = metrics
-        self.best_loss = np.inf
-        self.best_metric = {k: 0.0 for k in self.metric.keys()}
-        self.log_step = log_step
-        self.val_step = val_step
-        self.scaler = GradScaler(enabled=self.fp16)
-        self.scheduler = StepLR(self.optimizer, step_size=3, gamma=0.1)
-        for epoch in range(n_epochs):
-            self.print(f"\nEpoch {epoch:>3d}")
-            self.print("-----------------------------------")
 
-            # 1: Training phase
-            # 1.1 train
-            avg_loss = self.train_epoch(epoch=epoch, dataloader=self.train_data)
-
-            # 1.2 log result
-            self.print("+ Training result")
-            self.print(f"Loss: {avg_loss}")
-            for m in self.metric.values():
-                m.summary()
-
-            # 2: Evalutation phase
-            if (epoch + 1) % self.val_step == 0:
-                with autocast(enabled=self.fp16):
-                    # 2: Evaluating model
-                    avg_loss = self.evaluate(epoch, dataloader=self.val_data)
-
-                    self.print("+ Evaluation result")
-                    self.print(f"Loss: {avg_loss}")
-                    for m in self.metric.values():
-                        m.summary()
-
-                    # 3: Learning rate scheduling
-                    self.scheduler.step(avg_loss)
-
-                    # 4: Saving checkpoints
-                    if not self.debug:
-                        # Get latest val loss here
-                        val_metric = {k: m.value() for k, m in self.metric.items()}
-                        self.save_checkpoint(epoch, avg_loss, val_metric)
-            self.print("-----------------------------------")
+    def fit():
+        raise NotImplementedError
 
     def train_epoch(self, epoch: int, dataloader: DataLoader) -> float:
         running_loss = AverageValueMeter()
@@ -133,8 +51,8 @@ class SupervisedLearner(BaseLearner):
             self.optimizer.zero_grad()
             with autocast(enabled=self.fp16):
                 # 3: Get network outputs
-                outs, loss, loss_dict = self.model(batch)
                 # 4: Calculate the loss
+                outs, loss, loss_dict = self.model(batch)
             # 5: Calculate gradients
             self.scaler.scale(loss).backward()
             self.scaler.unscale_(self.optimizer)
@@ -162,45 +80,11 @@ class SupervisedLearner(BaseLearner):
         avg_loss = total_loss.value()[0]
         return avg_loss
 
-    def save_checkpoint(self, epoch, val_loss, val_metric):
-        data = {
-            "epoch": epoch,
-            "model_state_dict": self.model.state_dict(),
-            "optimizer_state_dict": self.optimizer.state_dict(),
-        }
+    def save_checkpoints():
+        raise NotImplementedError
 
-        if val_loss < self.best_loss:
-            self.print(
-                f"Loss is improved from {self.best_loss: .6f} to {val_loss: .6f}. Saving weights...",
-            )
-            save_model(data, self.save_dir / Path("best_loss.pth"))
-            # Update best_loss
-            self.best_loss = val_loss
-        else:
-            self.print(f"Loss is not improved from {self.best_loss:.6f}.")
-
-        for k in self.metric.keys():
-            if val_metric[k] > self.best_metric[k]:
-                self.print(
-                    f"{k} is improved from {self.best_metric[k]: .6f} to {val_metric[k]: .6f}. Saving weights...",
-                )
-                save_model(data, self.save_dir / Path(f"best_metric_{k}.pth"))
-                self.best_metric[k] = val_metric[k]
-            else:
-                self.print(f"{k} is not improved from {self.best_metric[k]:.6f}.")
-
-    @torch.no_grad()
-    def evaluate(self, epoch, dataloader):
-        avg_loss, metric = evaluate(
-            model=self.model,
-            dataloader=dataloader,
-            metric=self.metric,
-            device=self.device,
-            criterion=self.criterion,
-            verbose=self.verbose,
-        )
-        self.metric = metric
-        return avg_loss
+    def save_result():
+        raise NotImplementedError
 
     def print(self, obj):
         vprint(obj, self.verbose)
